Allow handling of utf-8 encoded files
Currently, if you try to register a word like "Walchaís" (not accent on i) with `language_tool_python`, so that it is ignored in spell check, we get a Unicode encoding error. This can be fixed by letting the functions which read/write custom spellings: `_register_spellings` and `_unregister_spellings`.

We could improve this by letting these functions take a variable called `encoding`.