fix: as a lib we should not configure the root logger
While setting the basicConfig value to the root logger it creates a new
StreamHandler and adds it to the root logger.
When an app that uses the lib language_tool_python also import the logging
lib, its own root logger get affected by this line and logs are printed
two times.